댓글 파일 업로드 기능 구현

+++ backend/src/main/java/org/example/backend/model/entity/board/Reply.java
@@ -29,6 +29,6 @@ public class Reply extends BaseTimeEntity2 {
     private Long replyId;       // 댓글 pk
     private Long boardId;       // 글번호 fk
     private String memberId;    // 댓글작성자 fk
-    private Integer reReply;    // 대댓글 fk
+    private Long reReply;    // 대댓글 fk
     private String reply;       // 내용
 }
